routing part of dsl slightly modified, simplified and beutified
